feat(domain): use cases for orders and not allowed error
Creation the entity and repository for orders.
Created use cases: place order, cancel order, mark as delivered, edit order, fetch all orders, fetch bag items, fetch open orders, history customer's
orders.
Created orders in-memory repository and factory for use cases testing.
Created error class for action not allowed.

+++ src/domain/commerce/application/use-cases/remove-from-bag.ts
@@ -1,6 +1,7 @@
 import { CustomersRepository } from '@account/application/repositories/customers-repository';
 import { BagItemsRepository } from '../repositories/bag-items-repository';
 import { ResourceNotFoundError } from '@account/application/use-cases/errors/resource-not-found-error';
+import { NotAllowedError } from '@account/application/use-cases/errors/not-allowed-error';
 
 interface RemoveFromBagUseCaseRequest {
   customerId: string;
@@ -31,7 +32,11 @@ export class RemoveFromBagUseCase {
     }
 
     if (customerId !== item.customerId) {
-      throw new Error('Not allowed.');
+      throw new NotAllowedError();
+    }
+
+    if (item.orderId) {
+      throw new NotAllowedError();
     }
 
     await this.bagItemsRepository.remove(itemId);
